backend/requirements.txt
Flask==2.3.3
werkzeug==2.3.7
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
python-dotenv==1.0.0
anthropic==0.7.0
python-dateutil==2.8.2
jsonschema==4.20.0
requests==2.31.0
pandas==2.1.4
scikit-learn==1.3.2
numpy==1.24.3
joblib==1.3.2
